refactor(review): move commit mechanics behind target-edit seam
