+++ packages/lexical-review/src/ReviewTargeting.ts
@@ -20,16 +20,12 @@ import {
   getTextChildren,
   isReviewElementNode,
   isRootParagraph,
-  type ReviewDeletionNode,
   type ReviewElementNode,
   type ReviewFragmentNode,
 } from "./ReviewNodes";
 import {
-  changed,
   refusal,
-  unchanged,
   type Preparation,
-  type ReviewIntentOutcome,
   type ReviewIntentRefusal,
 } from "./ReviewIntent";
 
@@ -74,7 +70,7 @@ type OffsetEntry = Readonly<{
   start: number;
 }>;
 
-type ProposalMapEntry = OffsetEntry &
+export type ProposalMapEntry = OffsetEntry &
   Readonly<{
     wrapper: ReviewElementNode;
   }>;
@@ -93,7 +89,7 @@ type ProposalSpan = Readonly<{
   start: number;
 }>;
 
-type AcceptedMapEntry = OffsetEntry &
+export type AcceptedMapEntry = OffsetEntry &
   Readonly<{
     childIndex: number;
   }>;
@@ -225,7 +221,7 @@ function validateSelectionFormatting(
     : null;
 }
 
-function isTextBoundary(text: string, offset: number): boolean {
+export function isTextBoundary(text: string, offset: number): boolean {
   if (offset <= 0 || offset >= text.length) {
     return true;
   }
@@ -852,7 +848,7 @@ function buildProposalMapAroundPoint(
   return buildProposalMap(point.paragraph, startWrapper, endWrapper);
 }
 
-function getProposalOffset(
+export function getProposalOffset(
   point: ProposalPoint,
   map: ProposalMap,
 ): number | null {
@@ -921,7 +917,7 @@ function buildProposalSpan(
   return { status: "ready", value: { end, map: map.value, start } };
 }
 
-function buildAcceptedMap(paragraph: ParagraphNode): AcceptedMap {
+export function buildAcceptedMap(paragraph: ParagraphNode): AcceptedMap {
   const children = paragraph.getChildren();
   const collected = collectOffsetUnits(
     children,
@@ -948,7 +944,7 @@ function buildAcceptedMap(paragraph: ParagraphNode): AcceptedMap {
   };
 }
 
-function getAcceptedOffset(
+export function getAcceptedOffset(
   point: AcceptedPoint,
   map: AcceptedMap,
 ): number | null {
@@ -1033,373 +1029,12 @@ function buildAcceptedSpan(
   return { status: "ready", value: { end, map, start } };
 }
 
-function getStartEntry(
-  entries: readonly ProposalMapEntry[],
-  offset: number,
-): ProposalMapEntry | null;
-function getStartEntry(
-  entries: readonly AcceptedMapEntry[],
-  offset: number,
-): AcceptedMapEntry | null;
-function getStartEntry(
-  entries: readonly AcceptedMapEntry[] | readonly ProposalMapEntry[],
-  offset: number,
-): AcceptedMapEntry | ProposalMapEntry | null {
-  return (
-    entries.find((entry) => entry.start <= offset && offset < entry.end) ?? null
-  );
-}
-
-function getEndEntry(
-  entries: readonly ProposalMapEntry[],
-  offset: number,
-): ProposalMapEntry | null;
-function getEndEntry(
-  entries: readonly AcceptedMapEntry[],
-  offset: number,
-): AcceptedMapEntry | null;
-function getEndEntry(
-  entries: readonly AcceptedMapEntry[] | readonly ProposalMapEntry[],
-  offset: number,
-): AcceptedMapEntry | ProposalMapEntry | null {
-  return (
-    entries.find((entry) => entry.start < offset && offset <= entry.end) ?? null
-  );
-}
-
 /**
- * Span-local edit operations behind the interaction-target seam. Owners
- * classify through inspectReviewTarget, then edit through these operations
- * instead of walking offset maps and entries themselves.
+ * Span-node access for the formatting owner behind the interaction-target
+ * seam. Commit mechanics (accepted-deletion math, proposal-deletion
+ * preparation, proposal mutation, caret restore) live in ReviewTargetEdit
+ * behind $commitTargetEdit.
  */
-
-export function isolateAcceptedTextRange(
-  target: AcceptedRangeTarget,
-): TextNode[] | null {
-  if (target.start >= target.end) {
-    return null;
-  }
-  const map = buildAcceptedMap(target.paragraph);
-  const startEntry = getStartEntry(map.entries, target.start);
-  const endEntry = getEndEntry(map.entries, target.end);
-  if (startEntry === null || endEntry === null) {
-    return null;
-  }
-  const startOffset = target.start - startEntry.start;
-  const endOffset = target.end - endEntry.start;
-  if (startEntry.node.getKey() === endEntry.node.getKey()) {
-    const parts = startEntry.node.splitText(startOffset, endOffset);
-    const selected = startOffset === 0 ? parts[0] : parts[1];
-    return selected === undefined ? null : [selected];
-  }
-
-  let first = startEntry.node;
-  if (startOffset > 0) {
-    const parts = first.splitText(startOffset);
-    first = parts[1] ?? first;
-  }
-  let last = endEntry.node;
-  if (endOffset < last.getTextContentSize()) {
-    const parts = last.splitText(endOffset);
-    last = parts[0] ?? last;
-  }
-  const firstIndex = getChildIndex(target.paragraph, first);
-  const lastIndex = getChildIndex(target.paragraph, last);
-  if (firstIndex === null || lastIndex === null || firstIndex > lastIndex) {
-    return null;
-  }
-  const selected = target.paragraph
-    .getChildren()
-    .slice(firstIndex, lastIndex + 1);
-  return selected.every($isTextNode) ? selected : null;
-}
-
-/**
- * Split-free read of an accepted range: joined span text plus whether every
- * overlapping node carries the live selection format. Null when the span
- * cannot be resolved in the live tree. Read-only: no splits, no caret moves.
- */
-export function readAcceptedRangeText(
-  target: AcceptedRangeTarget,
-): { text: string; uniformSelectionFormat: boolean } | null {
-  const map = buildAcceptedMap(target.paragraph);
-  const startEntry = getStartEntry(map.entries, target.start);
-  const endEntry = getEndEntry(map.entries, target.end);
-  if (startEntry === null || endEntry === null) {
-    return null;
-  }
-  let text = "";
-  let uniformSelectionFormat = true;
-  for (const entry of map.entries) {
-    if (entry.end <= target.start || entry.start >= target.end) {
-      continue;
-    }
-    const sliceStart = Math.max(target.start - entry.start, 0);
-    const sliceEnd = Math.min(
-      target.end - entry.start,
-      entry.end - entry.start,
-    );
-    text += entry.node.getTextContent().slice(sliceStart, sliceEnd);
-    if (entry.node.getFormat() !== target.selection.format) {
-      uniformSelectionFormat = false;
-    }
-  }
-  return { text, uniformSelectionFormat };
-}
-
-export function placeProposalCaret(
-  paragraph: ParagraphNode,
-  wrappers: readonly ReviewElementNode[],
-  offset: number,
-  fallbackIndex: number,
-): void {
-  let cursor = 0;
-  let lastText: TextNode | null = null;
-  for (const child of wrappers) {
-    // Removing wrappers can clone the paragraph; its key remains stable.
-    if (child.getParent()?.getKey() !== paragraph.getKey()) {
-      continue;
-    }
-    const textChildren = getTextChildren(child);
-    if (textChildren === null) {
-      continue;
-    }
-    for (const textNode of textChildren) {
-      const length = textNode.getTextContentSize();
-      if (offset <= cursor + length) {
-        textNode.select(offset - cursor, offset - cursor);
-        return;
-      }
-      cursor += length;
-      lastText = textNode;
-    }
-  }
-  if (lastText !== null) {
-    lastText.selectEnd();
-    return;
-  }
-  const paragraphOffset = Math.min(
-    Math.max(fallbackIndex, 0),
-    paragraph.getChildrenSize(),
-  );
-  paragraph.select(paragraphOffset, paragraphOffset);
-}
-
-export function spliceProposalRange(
-  target: ProposalRangeTarget | ProposalCaretTarget,
-  start: number,
-  end: number,
-  replacement: Readonly<{ node: TextNode; text: string }> | null = null,
-): Preparation<void> {
-  const mapped = proposalMapOf(target);
-  if (mapped.status !== "ready") return mapped;
-  const entries = mapped.value.entries;
-  for (let index = entries.length - 1; index >= 0; index -= 1) {
-    const entry = entries[index];
-    if (entry === undefined) {
-      continue;
-    }
-    const localStart = Math.max(start, entry.start) - entry.start;
-    const localEnd = Math.min(end, entry.end) - entry.start;
-    if (localStart >= localEnd) {
-      continue;
-    }
-    if (
-      replacement !== null &&
-      entry.node.getKey() === replacement.node.getKey()
-    ) {
-      entry.node.spliceText(
-        localStart,
-        localEnd - localStart,
-        replacement.text,
-        true,
-      );
-    } else {
-      entry.node.spliceText(localStart, localEnd - localStart, "", false);
-      if (entry.node.getTextContentSize() === 0) {
-        entry.node.remove();
-      }
-    }
-  }
-  for (const wrapper of target.wrappers) {
-    if (wrapper.getChildrenSize() === 0) {
-      wrapper.remove();
-    }
-  }
-  return { status: "ready", value: undefined };
-}
-
-export function replaceProposalRange(
-  target: ProposalRangeTarget,
-  text: string,
-): ReviewIntentOutcome {
-  if (target.start === target.end) {
-    return unchanged();
-  }
-  const mapped = proposalMapOf(target);
-  if (mapped.status !== "ready") return mapped;
-  const startEntry = getStartEntry(mapped.value.entries, target.start);
-  if (startEntry === null)
-    return refusal(
-      "invalid-structural-target",
-      "The proposal selection points cannot be resolved in the live tree.",
-    );
-  const fallbackIndex = getChildIndex(target.paragraph, target.wrappers[0]!);
-  const spliced = spliceProposalRange(target, target.start, target.end, {
-    node: startEntry.node,
-    text,
-  });
-  if (spliced.status !== "ready") return spliced;
-  placeProposalCaret(
-    target.paragraph,
-    target.wrappers,
-    target.start + text.length,
-    fallbackIndex ?? 0,
-  );
-  return changed();
-}
-
-function acceptedRange(
-  target: AcceptedCaretTarget,
-  backward: boolean,
-  start: number,
-  end: number,
-): AcceptedRangeTarget | null {
-  const map = buildAcceptedMap(target.paragraph);
-  const startEntry = getStartEntry(map.entries, start);
-  const endEntry = getEndEntry(map.entries, end);
-  if (startEntry === null || endEntry === null) {
-    return null;
-  }
-  return {
-    kind: "accepted-range",
-    paragraph: target.paragraph,
-    start,
-    end,
-    backward,
-    selection: target.selection,
-  };
-}
-
-export function acceptedDeletionTarget(
-  target: AcceptedCaretTarget,
-  backward: boolean,
-  granularity: "character" | "word",
-): AcceptedRangeTarget | null {
-  const map = buildAcceptedMap(target.paragraph);
-  const offset = getAcceptedOffset(
-    {
-      association: "accepted",
-      childIndex: target.childIndex,
-      node: target.node,
-      offset: target.offset,
-      paragraph: target.paragraph,
-    },
-    map,
-  );
-  if (offset === null) {
-    return null;
-  }
-  if (granularity === "word") {
-    const children = target.paragraph.getChildren();
-    let index = target.childIndex;
-    if (target.node === null && backward) index -= 1;
-    if (!$isTextNode(children[index])) return null;
-    let left = index;
-    let right = index;
-    while ($isTextNode(children[left - 1])) left -= 1;
-    while ($isTextNode(children[right + 1])) right += 1;
-    const run = map.entries.filter(
-      (entry) => entry.childIndex >= left && entry.childIndex <= right,
-    );
-    const base = run[0]?.start;
-    if (base === undefined) return null;
-    const text = run.map((entry) => entry.node.getTextContent()).join("");
-    const local = offset - base;
-    const boundary = deletionOffset(text, local, backward, granularity);
-    return acceptedRange(
-      target,
-      backward,
-      base + Math.min(local, boundary),
-      base + Math.max(local, boundary),
-    );
-  }
-  if (target.node !== null) {
-    const text = target.node.getTextContent();
-    if (backward && target.offset > 0) {
-      return acceptedRange(
-        target,
-        backward,
-        offset - (target.offset - previousCharacterOffset(text, target.offset)),
-        offset,
-      );
-    }
-    if (!backward && target.offset < text.length) {
-      return acceptedRange(
-        target,
-        backward,
-        offset,
-        offset + (nextCharacterOffset(text, target.offset) - target.offset),
-      );
-    }
-  }
-  const adjacentIndex = backward
-    ? target.childIndex - 1
-    : target.node === null
-      ? target.childIndex
-      : target.childIndex + 1;
-  const adjacent = target.paragraph.getChildAtIndex(adjacentIndex);
-  if (!$isTextNode(adjacent) || adjacent.getTextContentSize() === 0) {
-    return null;
-  }
-  const adjacentEntry = map.entries.find(
-    (entry) => entry.node.getKey() === adjacent.getKey(),
-  );
-  if (adjacentEntry === undefined) {
-    return null;
-  }
-  if (backward) {
-    const start = previousCharacterOffset(
-      adjacent.getTextContent(),
-      adjacent.getTextContentSize(),
-    );
-    return acceptedRange(
-      target,
-      backward,
-      adjacentEntry.start + start,
-      adjacentEntry.end,
-    );
-  }
-  const end = nextCharacterOffset(adjacent.getTextContent(), 0);
-  return acceptedRange(
-    target,
-    backward,
-    adjacentEntry.start,
-    adjacentEntry.start + end,
-  );
-}
-
-function deletionOffset(
-  text: string,
-  offset: number,
-  backward: boolean,
-  granularity: "character" | "word",
-): number {
-  if (granularity === "character")
-    return backward
-      ? previousCharacterOffset(text, offset)
-      : nextCharacterOffset(text, offset);
-  // A word intention consumes adjacent whitespace followed by one word or
-  // punctuation run. Boundaries are computed without changing live selection.
-  const pattern = backward
-    ? /(?:[\p{L}\p{N}\p{M}_]+|[^\p{L}\p{N}\p{M}_\s]+)\s*$|\s+$/u
-    : /^\s*(?:[\p{L}\p{N}\p{M}_]+|[^\p{L}\p{N}\p{M}_\s]+)|^\s+/u;
-  const match = (backward ? text.slice(0, offset) : text.slice(offset)).match(
-    pattern,
-  );
-  const length = match?.[0].length ?? 0;
-  return backward ? offset - length : offset + length;
-}
 
 /** Identity-axis classification behind the seam: kind only, no live nodes leak. */
 export type ProposalKind =
@@ -1411,119 +1046,6 @@ export function inspectProposalKind(
   const group = inspectProposalGroup(proposalId);
   if (group.status !== "ready") return group;
   return { status: "ready", value: group.value.kind };
-}
-
-function isReplacementKind(proposalId: string): boolean {
-  const kind = inspectProposalKind(proposalId);
-  return kind.status === "ready" && kind.value === "replacement";
-}
-
-export type ProposalCaretDeletion =
-  | Readonly<{ action: "splice"; start: number; end: number }>
-  | Readonly<{ action: "resolve-deletion" }>
-  | Readonly<{ action: "resolve-replacement" }>;
-
-/**
- * Proposal-caret deletion math behind the seam. Resolve actions name the
- * owner-side resolution the caller performs; this module never imports it.
- */
-export function prepareProposalCaretDeletion(
-  target: ProposalCaretTarget,
-  backward: boolean,
-  granularity: "character" | "word",
-): Preparation<ProposalCaretDeletion> {
-  if ($isReviewFormattingNode(target.wrapper))
-    return refusal(
-      "unsupported-proposal-edit",
-      "Text deletion cannot alter a pending formatting target.",
-    );
-  const mapped = proposalMapOf(target);
-  if (mapped.status !== "ready") return mapped;
-  const offset = getProposalOffset(
-    {
-      association: "proposal",
-      childIndex: target.childIndex,
-      node: target.node,
-      offset: target.offset,
-      paragraph: target.paragraph,
-      wrapper: target.wrapper,
-    },
-    mapped.value,
-  );
-  if (offset === null)
-    return refusal(
-      "invalid-structural-target",
-      "The proposal caret cannot be resolved in the live tree.",
-    );
-  const text = mapped.value.entries
-    .map((entry) => entry.node.getTextContent())
-    .join("");
-  if (!isTextBoundary(text, offset)) {
-    return refusal(
-      "invalid-structural-target",
-      "The proposal caret is not on a supported Unicode text boundary.",
-    );
-  }
-  if (backward && offset === 0) {
-    return refusal(
-      "deletion-target-unavailable",
-      "Backward deletion may not cross from proposal content into accepted content.",
-    );
-  }
-  if (!backward && offset === mapped.value.total) {
-    return refusal(
-      "deletion-target-unavailable",
-      "Forward deletion may not cross from proposal content into accepted content.",
-    );
-  }
-  if ($isReviewDeletionNode(target.wrapper))
-    return { status: "ready", value: { action: "resolve-deletion" } };
-  const boundary = deletionOffset(text, offset, backward, granularity);
-  const start = Math.min(offset, boundary);
-  const end = Math.max(offset, boundary);
-  if (
-    end - start === mapped.value.total &&
-    isReplacementKind(target.proposalId)
-  )
-    return { status: "ready", value: { action: "resolve-replacement" } };
-  return { status: "ready", value: { action: "splice", start, end } };
-}
-
-export type ProposalRangeDeletion =
-  | Readonly<{ action: "splice" }>
-  | Readonly<{ action: "unchanged" }>
-  | Readonly<{ action: "resolve-deletion" }>
-  | Readonly<{ action: "resolve-replacement" }>;
-
-/**
- * Proposal-range deletion behind the seam. Check order mirrors the former
- * owner logic exactly: formatting, empty span, deletion, validation,
- * replacement. Empty spans resolve here so resolve branches keep their
- * original position on both sides of the empty check.
- */
-export function prepareProposalRangeDeletion(
-  target: ProposalRangeTarget,
-): Preparation<ProposalRangeDeletion> {
-  if ($isReviewFormattingNode(target.wrappers[0]))
-    return refusal(
-      "unsupported-proposal-edit",
-      "Text deletion cannot alter a pending formatting target.",
-    );
-  if (target.start === target.end)
-    return { status: "ready", value: { action: "unchanged" } };
-  if ($isReviewDeletionNode(target.wrappers[0]))
-    return { status: "ready", value: { action: "resolve-deletion" } };
-  const group = inspectProposalGroup(target.proposalId);
-  if (group.status !== "ready") return group;
-  const insertionTotal = group.value.wrappers
-    .filter($isReviewInsertionNode)
-    .reduce((sum, node) => sum + node.getTextContentSize(), 0);
-  if (
-    group.value.kind === "replacement" &&
-    target.end - target.start === insertionTotal
-  )
-    return { status: "ready", value: { action: "resolve-replacement" } };
-  return { status: "ready", value: { action: "splice" } };
 }
 
 /**
@@ -1652,7 +1174,7 @@ export function inspectFragmentSelection(): Preparation<FragmentSelection> | nul
  * Fresh proposal map for one target at use time. Owners never carry maps;
  * every span operation rebuilds from the target's paragraph and wrappers.
  */
-function proposalMapOf(target: {
+export function proposalMapOf(target: {
   paragraph: ParagraphNode;
   wrappers: readonly ReviewElementNode[];
 }): Preparation<ProposalMap> {
@@ -1664,44 +1186,6 @@ function proposalMapOf(target: {
       "The pending proposal has no live content.",
     );
   return buildProposalMap(target.paragraph, first, last);
-}
-
-/**
- * Accepted-deletion continuation behind the seam: edge nodes stay inside.
- */
-export function findAcceptedDeletionContinuation(
-  target: AcceptedRangeTarget,
-  backward: boolean,
-): Preparation<{
-  proposalId: string | null;
-  node: ReviewDeletionNode | null;
-}> {
-  const map = buildAcceptedMap(target.paragraph);
-  const first = getStartEntry(map.entries, target.start);
-  const last = getEndEntry(map.entries, target.end);
-  if (first === null || last === null)
-    return refusal(
-      "invalid-structural-target",
-      "The accepted selection points cannot be resolved in the live tree.",
-    );
-  const adjacent = backward
-    ? target.end === last.end
-      ? last.node.getNextSibling()
-      : null
-    : target.start === first.start
-      ? first.node.getPreviousSibling()
-      : null;
-  const continuation = $isReviewDeletionNode(adjacent) ? adjacent : null;
-  if (continuation === null)
-    return { status: "ready", value: { proposalId: null, node: null } };
-  const kind = inspectProposalKind(continuation.getProposalId());
-  if (kind.status !== "ready") return kind;
-  if (kind.value === "replacement")
-    return { status: "ready", value: { proposalId: null, node: null } };
-  return {
-    status: "ready",
-    value: { proposalId: continuation.getProposalId(), node: continuation },
-  };
 }
 
 export type StructuralPosition = Readonly<{
